moving sprites around was being done all wrong!
Using the sprite rect to track the position of entities meant that
sub-pixel positioning was being lost every frame. This meant projectiles
would gradually go off-course and player motion could be jittery. The
problem is easily fixed by storing the ideal position of the entity in a
separate vector and moving the image to align with that vector's
position each fram. This whole mess is an example of why it's a good
idea to keep game logic and presentation logic separate.